Make C stubs more resistant to header structure
The existing stubs cause a warning on OCaml 5.1 because the declaration
of `caml_convert_signal_number` can not be found.

The reason for that is that `<caml/signals.h>` is already included by
some headers at the beginning of the file, so `CAML_SIGNALS_H` is
already set by the time `<caml/signals.h>` is included manually and the
extra internal definition are not included.

This used to work in previous versions of OCaml because presumably the
header structure was different and `<caml/signal.h>` was not included
transitively.

This fix just sets `CAML_INTERNALS` at the file level; it is not really
possible to be granular in setting it.

Signed-off-by: Etienne Millon <me@emillon.org>